rename wing.py
