+++ python_code/3d/symmetric_5_sided_mesh.py
